requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
requests>=2.31.0
mysqlclient>=2.2; platform_system != "Windows"
# Windows 若无 mysqlclient 可改用: pip install pymysql 并在 __init__ 中 install_as_MySQLdb()
django-redis>=5.4
redis>=5.0
captcha>=0.5
Pillow>=10
openai>=1.12.0
langchain>=0.2.0
langchain-openai>=0.1.0
chromadb>=0.5.0
unstructured>=0.14.0
jsonpath-ng>=1.6.0
cryptography>=42.0.0
apscheduler>=3.10.4
django-apscheduler>=0.7.0
celery>=5.4
channels>=4.0
channels-redis>=4.2
daphne>=4.0
pdfplumber>=0.11.0
python-docx>=1.1.2
